refactor to use click
