+++ .dockerignore
@@ -1,0 +1,3 @@
+.idea/*
+env/*
+*.pyc
