+++ .gitignore
@@ -191,3 +191,4 @@ ENV/
 /site
 
 # End of https://www.gitignore.io/api/osx,python,intellij
+.idea/
